MicroscopeAOCompositeDevice: split GUI into separate class.
Move the whole UI into a separate class creating passthrough functions
in the device class as required.  Also add loads of new functions to
reduce code duplication.  Save data of what was done in the log
directory instead of the config directory.  Use the panel as parent
for the other UI so they inherit as expected.  Remove unused
attributes and methods.