AdaptiveOpticsDevice: set software trigger before multiple acquires (#18)
The calibrate, flatten_phase, and assess_character methods involve
multiple acquires.  Currently, if the camera is not set to software
trigger that is done right before acquire and then set back to what it
was before which in many cameras means disable and then reenable the
camera which takes a lot of time.

This commit changes the code from a contextmanager to a decorator and
decorates the methods that will require acquisition of multiple
images, i.e., calibrate, flatten_phase, and assess_character to do it
earlier.